+++ pydidas/gui/frames/pyfai_calib_frame.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2021-, Helmholtz-Zentrum Hereon
+# Copyright 2023, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 #
-# Parts of this file are adapted based on the pyfai-calib widget which
+# Parts of this file are adapted based on the pyFAI.gui.CalibrationWindow widget which
 # is distributed under the MIT license.
 
 """
@@ -24,7 +24,7 @@ to be used within pydidas.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
@@ -40,13 +40,6 @@ from pyFAI.app import calib2
 from pyFAI.gui.CalibrationContext import CalibrationContext
 from pyFAI.gui.CalibrationWindow import MenuItem
 from pyFAI.gui.model import MarkerModel
-from pyFAI.gui.tasks import (
-    ExperimentTask,
-    GeometryTask,
-    IntegrationTask,
-    MaskTask,
-    PeakPickingTask,
-)
 from pyFAI.gui.utils import projecturl
 from qtpy import QtCore, QtGui, QtWidgets
 from silx.gui.plot.tools import ImageToolBar
@@ -75,6 +68,14 @@ def _create_calib_tasks():
         The list with the tasks.
 
     """
+    from pyFAI.gui.tasks import (
+        ExperimentTask,
+        GeometryTask,
+        IntegrationTask,
+        MaskTask,
+        PeakPickingTask,
+    )
+
     tasks = [
         ExperimentTask.ExperimentTask(),
         MaskTask.MaskTask(),